+++ gDRplots/NEWS.md
@@ -1,5 +1,8 @@
-## gDRplots 0.0.81 - 2025-06-06
+## gDRplots 0.0.82 - 2025-06-06
 * update `pheatmap_qc` with standardized concentrations
+
+## gDRplots 0.0.81 - 2025-05-28
+* update arg names in cgs functions
 
 ## gDRplots 0.0.80 - 2025-05-27
 * update chunk naming to fulfill `knitr` requirements